starting on main logic

+++ irsensor.h
@@ -2,5 +2,5 @@
 #define IRSENSOR_H
 
 void setup_irsensor(uint8_t PIN);
-float read_irsensor(uint8_t PIN);
+int read_irsensor(uint8_t PIN);
 #endif
